- all icolib services removed; - IcoLib_* functions are bound directly to the code
git-svn-id: http://svn.miranda-ng.org/main/trunk@14161 1316c22d-e87f-b044-9b9b-93d7a3e3ba9c

+++ include/m_core.h
@@ -699,7 +699,7 @@ public:
 		return res;
 	}
 
-	__forceinline char& operator[](int idx) const { return m_str[idx]; }
+	__forceinline char& operator[](size_t idx) const { return m_str[idx]; }
 	__forceinline operator char*() const {	return m_str; }
 	__forceinline operator unsigned char*() const {	return (unsigned char*)m_str; }
 	__forceinline operator LPARAM() const { return (LPARAM)m_str; }
